pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mercury"
version = "0.1.0-dev"
description = "Mercury — local-first dual-brain agent built on a fork of NousResearch/hermes-agent."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Red Team Kitchen" },
  { name = "Nous Research", email = "mercury@nousresearch.com" },
]
license = { text = "MIT" }
dependencies = [
  # Core — pinned to known-good ranges to limit supply chain attack surface
  "openai>=2.21.0,<3",
  "anthropic>=0.39.0,<1",
  "python-dotenv>=1.2.1,<2",
  "fire>=0.7.1,<1",
  "httpx[socks]>=0.28.1,<1",
  "rich>=14.3.3,<15",
  "tenacity>=9.1.4,<10",
  "pyyaml>=6.0.2,<7",
  "requests>=2.33.0,<3",  # CVE-2026-25645
  "jinja2>=3.1.5,<4",
  "pydantic>=2.12.5,<3",
  # Interactive CLI (prompt_toolkit is used directly by cli.py)
  "prompt_toolkit>=3.0.52,<4",
  # Tools
  "exa-py>=2.9.0,<3",
  "firecrawl-py>=4.16.0,<5",
  "parallel-web>=0.4.2,<1",
  "fal-client>=0.13.1,<1",
  # Text-to-speech (Edge TTS is free, no API key needed)
  "edge-tts>=7.2.7,<8",
  # Skills Hub (GitHub App JWT auth — optional, only needed for bot identity)
  "PyJWT[crypto]>=2.12.0,<3",  # CVE-2026-32597
]

[project.optional-dependencies]
dev = ["debugpy>=1.8.0,<2", "pytest>=9.0.2,<10", "pytest-asyncio>=1.3.0,<2", "pytest-xdist>=3.0,<4", "mcp>=1.2.0,<2", "ty>=0.0.1a29,<0.0.22", "ruff"]
messaging = ["discord.py[voice]>=2.7.1,<3", "aiohttp>=3.13.3,<4", "qrcode>=7.0,<8"]
cron = ["croniter>=6.0.0,<7"]
cli = ["simple-term-menu>=1.0,<2"]
tts-premium = ["elevenlabs>=1.0,<2"]
voice = [
  # Local STT pulls in wheel-only transitive deps (ctranslate2, onnxruntime),
  # so keep it out of the base install for source-build packagers like Homebrew.
  "faster-whisper>=1.0.0,<2",
  "sounddevice>=0.4.6,<1",
  "numpy>=1.24.0,<3",
]
pty = [
  "ptyprocess>=0.7.0,<1; sys_platform != 'win32'",
  "pywinpty>=2.0.0,<3; sys_platform == 'win32'",
]
mcp = ["mcp>=1.2.0,<2"]
google = [
  "google-api-python-client>=2.100,<3",
  "google-auth-oauthlib>=1.0,<2",
  "google-auth-httplib2>=0.2,<1",
  # Unified Gen AI SDK — same SDK for AI Studio (free dev key) and Vertex
  # AI / Gemini Enterprise Agent Platform.  Required by Mercury's Gemini
  # adapter; flip vertexai=True at call site for Vertex.
  "google-genai>=1.0.0,<2",
]
# `mercury dashboard` (localhost SPA + API).  Not in core to keep the
# default install lean — but always installed for Mercury since the
# Tailscale-only WebUI is a primary surface.
web = ["fastapi>=0.104.0,<1", "uvicorn[standard]>=0.24.0,<1"]
rl = [
  "atroposlib @ git+https://github.com/NousResearch/atropos.git@c20c85256e5a45ad31edf8b7276e9c5ee1995a30",
  "tinker @ git+https://github.com/thinking-machines-lab/tinker.git@30517b667f18a3dfb7ef33fb56cf686d5820ba2b",
  "fastapi>=0.104.0,<1",
  "uvicorn[standard]>=0.24.0,<1",
  "wandb>=0.15.0,<1",
]
yc-bench = ["yc-bench @ git+https://github.com/collinear-ai/yc-bench.git@bfb0c88062450f46341bd9a5298903fc2e952a5c ; python_version >= '3.12'"]
all = [
  "mercury[messaging]",
  "mercury[cron]",
  "mercury[cli]",
  "mercury[dev]",
  "mercury[tts-premium]",
  "mercury[pty]",
  "mercury[mcp]",
  "mercury[voice]",
  "mercury[google]",
  "mercury[web]",
]

[project.scripts]
mercury = "mercury_cli.main:main"
mercury-agent = "run_agent:main"

[tool.setuptools]
py-modules = ["run_agent", "model_tools", "toolsets", "batch_runner", "trajectory_compressor", "toolset_distributions", "cli", "mercury_constants", "mercury_state", "mercury_time", "mercury_logging", "rl_cli", "utils"]

[tool.setuptools.package-data]
mercury_cli = ["web_dist/**/*"]

[tool.setuptools.packages.find]
include = ["agent", "agent.*", "tools", "tools.*", "mercury_cli", "gateway", "gateway.*", "tui_gateway", "tui_gateway.*", "cron", "plugins", "plugins.*", "mercury", "mercury.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (API keys, Modal, etc.)",
]
addopts = "-m 'not integration' -n auto"

[tool.ty.environment]
python-version = "3.13"

[tool.ty.rules]
unknown-argument = "warn"
redundant-cast = "ignore"

[tool.ty.src]
exclude = ["**"]

[[tool.ty.overrides]]
include = ["**"]

[tool.ty.overrides.rules]
unresolved-import = "ignore"
invalid-method-override = "ignore"
invalid-assignment = "ignore"
not-iterable = "ignore"

[tool.ruff]
exclude = ["*"]

[tool.uv]
exclude-newer = "7 days"
